Fix Windows UTF-8 writes; tighten MCP builder types; fix tools test collection
