--- tools/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Aden Tools test suite."""
