--- tools/tests/tools/__init__.py
@@ -1,1 +1,0 @@
-"""Tool-specific tests."""
